+++ app/src/androidTest/java/woowacourse/movie/view/movies/MovieListActivityTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.movie
+package woowacourse.movie.view.movies
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
