+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
             android:name=".view.booking.BookingCompleteActivity"
             android:exported="false" />
         <activity
-            android:name=".view.movie.MovieListActivity"
+            android:name=".view.movies.MovieListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
