+++ app/src/main/res/layout/activity_movie_list.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.movie.MovieListActivity">
+    tools:context=".view.movies.MovieListActivity">
 
     <ListView
         android:id="@+id/list_view"
